transformer/requirements.txt
tensorboardX
torchtext==0.6.0
tqdm
spacy
